Refactor code, moved app file, templates & test to todo_app folder.
