+++ todo_app/card.py
@@ -1,0 +1,10 @@
+
+class Card:
+    def __init__(self, id, name, status, mod_date):
+        self.id = id
+        self.status = status
+        self.card_name = name
+        self.mod_date = mod_date        
+
+
+
